+++ dart/geobase/lib/src/coordinates/scalable/scalable2i.dart
@@ -211,15 +211,15 @@ class Scalable2i implements Scalable, Projected {
   @override
   bool equals2D(
     Position other, {
-    num toleranceHoriz = doublePrecisionEpsilon,
+    double toleranceHoriz = doublePrecisionEpsilon,
   }) =>
       Position.testEquals2D(this, other, toleranceHoriz: toleranceHoriz);
 
   @override
   bool equals3D(
     Position other, {
-    num toleranceHoriz = doublePrecisionEpsilon,
-    num toleranceVert = doublePrecisionEpsilon,
+    double toleranceHoriz = doublePrecisionEpsilon,
+    double toleranceVert = doublePrecisionEpsilon,
   }) =>
       Position.testEquals3D(
         this,